+++ lib/bmg/database/sequel.rb
@@ -7,13 +7,13 @@ module Bmg
 
       def initialize(sequel_db, options = {})
         @sequel_db = sequel_db
+        @sequel_db = ::Sequel.connect(@sequel_db) unless @sequel_db.is_a?(::Sequel::Database)
       end
 
       def method_missing(name, *args, &bl)
         return super(name, *args, &bl) unless args.empty? && bl.nil?
         raise NotSuchRelationError(name.to_s) unless @sequel_db.table_exists?(name)
-        table = @sequel_db[name]
-        rel_for(table)
+        rel_for(name)
       end
 
       def each_relation_pair